Speech_to_Speech_API/requirements.txt
numpy
openai
pyaudio
pygame
soundfile
python-dotenv
